requirements.txt
# Wrapper for Bugzilla API
python-bugzilla==3.2.0 

# Progress bars
tqdm==4.63.0